# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.AvailabilityResponse:
    properties:
      available:
        type: boolean
      currency:
        type: string
      nightly_rates:
        items:
          $ref: '#/definitions/service.NightlyRate'
        type: array
      nights:
        type: integer
      total_price:
        type: number
    type: object
  handler.BookingCancelResponse:
    properties:
      booking:
        $ref: '#/definitions/models.Booking'
      payment:
        $ref: '#/definitions/models.Payment'
    type: object
...
  handler.CreateBookingRequest:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      property_id:
        type: string
      room_id:
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handler.PayBookingRequest:
    properties:
      provider:
        type: string
      reference:
        type: string
    type: object
  handler.PaymentInvoiceResponse:
    properties:
      invoice:
        $ref: '#/definitions/models.Invoice'
      payment:
        $ref: '#/definitions/models.Payment'
    type: object
...
  handler.RegisterGuestRequest:
    properties:
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        $ref: '#/definitions/models.Gender'
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  handler.TokenResponseDoc:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        additionalProperties: true
        type: object
    type: object
...
  models.Booking:
    properties:
      booking_status:
        $ref: '#/definitions/models.BookingStatus'
      check_in:
        type: string
      check_out:
        type: string
      created_at:
        type: string
      guest_id:
        type: string
      id:
        type: string
      nights:
        type: integer
      note:
        type: string
      property_id:
        type: string
      refund_amount:
        type: number
      room_id:
        type: string
      total_price:
        type: number
    type: object
  models.BookingStatus:
    enum:
    - New
    - Confirmed
    - Cancelled
    - CheckedIn
    - CheckedOut
    - NoShow
    type: string
    x-enum-varnames:
    - BookingStatusNew
    - BookingStatusConfirmed
    - BookingStatusCancel
    - BookingStatusCheckedIn
    - BookingStatusCheckedOut
    - BookingStatusNoShow
  models.Gender:
    enum:
    - Male
    - Female
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
  models.Guest:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        $ref: '#/definitions/models.Gender'
      guest_type:
        $ref: '#/definitions/models.GuestType'
      id:
        type: string
      last_name:
        type: string
      nationality:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      vip_status:
        $ref: '#/definitions/models.VIPStatus'
    type: object
  models.GuestType:
    enum:
    - Adult
    - Child
    type: string
    x-enum-varnames:
    - GuestTypeAdult
    - GuestTypeChild
...
  models.Invoice:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      id:
        type: string
      invoice_number:
        type: string
      issued_at:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
    type: object
  models.Payment:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      paid_at:
        type: string
      provider:
        type: string
      reference:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
    type: object
  models.PaymentStatus:
    enum:
    - Pending
    - Paid
    - Refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusPaid
    - PaymentStatusRefunded
  models.Properties:
    properties:
      address:
        type: string
      auth_code:
        type: string
      cancellation_policy:
        type: string
      checkin_time:
        type: string
      checkout_time:
        type: string
      city:
        type: string
      created_at:
        type: string
      facilities:
        items:
          type: string
        type: array
      hotel_code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.PropertyDetailResponse:
    properties:
      property:
        $ref: '#/definitions/models.Properties'
      room_types:
        items:
          $ref: '#/definitions/models.RoomType'
        type: array
    type: object
...
  models.RoomType:
    properties:
      base_price:
        type: number
      capacity:
        type: integer
      created_at:
        type: string
      description:
        type: string
      facilities:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      property_id:
        type: string
    type: object
  models.VIPStatus:
    enum:
    - Bronze
    - Silver
    - Gold
    - Platinum
    type: string
    x-enum-varnames:
    - VIPStatusBronze
    - VIPStatusSilver
    - VIPStatusGold
    - VIPStatusPlatinum
  service.BookingCreateResult:
    properties:
      booking:
        $ref: '#/definitions/models.Booking'
      invoice:
        $ref: '#/definitions/models.Invoice'
      payment:
        $ref: '#/definitions/models.Payment'
      quote:
        $ref: '#/definitions/service.BookingQuote'
    type: object
  service.BookingQuote:
    properties:
      available:
        type: boolean
      currency:
        type: string
      nightly_rates:
        items:
          $ref: '#/definitions/service.NightlyRate'
        type: array
      nights:
        type: integer
      total_price:
        type: number
    type: object
  service.NightlyRate:
    properties:
      date:
        type: string
      rate:
        type: number
    type: object
...
info:
  contact: {}
  description: REST API for hotel booking management (guest, booking, admin inventory,
    reports).
  title: Hotel Booking API
  version: "1.0"
paths:
...
  /auth/guest/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload (email or phone)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TokenResponseDoc'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login guest
      tags:
      - Auth
  /auth/guest/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register guest
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterGuestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Guest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register guest
      tags:
      - Auth
  /guests/bookings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my bookings
      tags:
      - Guests
    post:
      consumes:
      - application/json
      parameters:
      - description: Create booking
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.BookingCreateResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - Guests
  /guests/bookings/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BookingCancelResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel booking
      tags:
      - Guests
  /guests/bookings/{id}/invoice:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Invoice'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get booking invoice
      tags:
      - Guests
  /guests/bookings/{id}/pay:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.PayBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.PaymentInvoiceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pay booking
      tags:
      - Guests
  /guests/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guest'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - Guests
  /hotels:
    get:
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Properties'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search hotels by city
      tags:
      - Hotels
  /hotels/{id}:
    get:
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PropertyDetailResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get hotel detail
      tags:
      - Hotels
  /rooms/{room_id}/availability:
    get:
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: string
      - description: Check-in date (YYYY-MM-DD)
        in: query
        name: check_in
        required: true
        type: string
      - description: Check-out date (YYYY-MM-DD)
        in: query
        name: check_out
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AvailabilityResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check room availability
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"